+++ src/pages/admin/Visitors.tsx
@@ -1,35 +1,14 @@
-import { useEffect, useState } from 'react';
 import AdminEntryCard from '../../components/common/card/admin/adminEntryCard/AdminEntryCard';
 import ResponsiveLayout from '../../components/common/layout/ResponsiveLayout';
 import * as S from './Visitors.style.ts';
-import { getConferenceInfo } from '../../api/reserve/getConferenceInfo.ts';
-
-export interface ConferenceInfo {
-  id: number;
-  name: string;
-  capacity: number;
-  startTime: string;
-  attend: number;
-  sessions: {
-    id: number;
-    name: string;
-    capacity: string;
-    startTime: string;
-  }[];
-}
+import useConferenceData from '../../hooks/useConferenceData.ts';
 
 const Visitors = () => {
-  const [conferInfo, setConferInfo] = useState<ConferenceInfo | null>(null);
+  const { conferenceDataQuery } = useConferenceData();
 
-  useEffect(() => {
-    const fetchConference = async () => {
-      const response = await getConferenceInfo(1);
-      if (response) {
-        setConferInfo(response.data);
-      }
-    };
-    fetchConference();
-  }, []);
+  if (conferenceDataQuery.isLoading) {
+    return <div>로딩 중...</div>;
+  }
 
   return (
     <ResponsiveLayout>
@@ -39,7 +18,7 @@ const Visitors = () => {
           해당 화면에 얼굴 인식으로 입장한 방문자 수가 표시돼요
         </S.Description>
       </S.TitleContainer>
-      <AdminEntryCard conferInfo={conferInfo} />
+      <AdminEntryCard conferInfo={conferenceDataQuery.data ?? null} />
     </ResponsiveLayout>
   );
 };
